gb/ppu: the mode 3 lead costs a quarter of what it did, and the rest is a floor
Eight lines of code; the rest is documentation. Zero behaviour change — all
three results files differ from the previous commit only in the timestamp,
gbgate is 26/26 identical over the real ROMs, and the blargg canary is 11/11
pixel-identical to SameBoy at frame 1200.

Retired instructions against 151b952, min of >=4 runs with cycles= matched:
Link's Awakening -0.600%, Pokemon Crystal -0.651%, Pokemon Blue -0.673%.
Against 93f7cc2, the commit before the fetch-phase fix: Zelda +0.926% ->
+0.283%, Crystal +0.952% -> +0.316%. So the accuracy work's cost drops from
about 1.2% to about 0.3%.

The profile. Building each tree with -d:M3_PIPE_DELAY=0 compiles the lead
mechanism out entirely and splits the bill:

                              lead machinery   everything else
  151b952  Link's Awakening        +0.831%          +0.095%
  151b952  Crystal                 +0.723%          +0.227%
  this     Link's Awakening        +0.224%          +0.060%
  this     Crystal                 +0.119%          +0.197%

The halves are additive to within 0.01 points on both titles. It was all the
lead, and the lead is arrangement rather than work — the pipeline runs the same
160 dots a line either way. Turning CGB_SCX/SCY_LATENCY from 0 to 2 in the same
commit, which switches on a whole write pipeline in memory.nim plus four new
out-of-line functions, is +0.01%: free.

Three hypotheses killed, two of them the obvious ones:

  * The inline cliff. A per-function size diff between lead-off and lead-on has
    exactly two rows, fifo_tick_slow and fifo_burst_tail. No opcode body moved.
    The 52-opcode cliff had already been mitigated and was not the residual.
  * fetcher_retired's four extra branches, +0.04%. Splitting them into a
    {.noinline.} tail behind the first compare — the shape that should make it
    one instruction on the dot loop — measures WORSE (+0.07%), because clang
    already folds that compare into the dot loop's own lx test (cmp w8,#0x9d /
    b.gt, one branch answering both questions) and the split takes that away.
  * {.noinline.} on the tail burst is no longer right. Inlining it costs
    fifo_tick_slow 172 bytes, changes the size of nothing else in the binary,
    and is worth -0.11%.

What actually changed: the head-delay block loses its `continue` and falls into
the dot loop (equivalent — the loop's own remaining > 0 exits when the head ate
the tick, and fetcher_retired cannot move across dots on which nothing ran);
fifo_burst_tail becomes {.inline.}; and m3_delay becomes uint8 so the
once-per-mode-3-M-cycle test is ldrb+cbz. Isolated those three sum to 0.22%;
together they measure -0.63%, a real non-additive codegen effect, verified via
the size diff not to be the cliff.

The floor. Of the +0.22% the lead still costs on DMG, about 0.19% is that
ldrb+cbz alone, roughly 6200 mode-3 M-cycles a frame. Four head spellings
against a build with the head removed outright: this one +0.162%, 151b952's
+0.187%, a {.noinline.} call +0.255%, a two-dot loop +0.318%. Deleting the test
would mean spending the head dots at the mode 2->3 edge, which is impossible: a
double-speed M-cycle is two dots and the transition is one of them, so a lead of
2 always leaves a dot for the next tick.

Not taken: making the head dots real pipeline dots (2 sleep steps in the first
fetch) is +0.6%, five times worse. The remaining non-lead +0.06%/+0.20% is the
fetch-phase fix's own behaviour (tick_bg_fetcher's hot part 268 -> 656 bytes)
and recovering it means changing what the fetcher does, which is out of scope
for a zero-behaviour-change commit.

Two corrections to docs/gb_oam_dma_cost.md, both of which produced wrong-signed
answers before they were caught:

  * ri_instructions is LOAD-SENSITIVE. Five runs of one binary spread 0.49% at
    load average ~100, correlated with each run's own fps, because kernel work
    is charged to the process. The "reproduces to 0.002%" figure in that doc is
    an idle-machine number and has been cited as unqualified all through this
    work. Take the minimum of at least four runs.
  * Two builds of identical source in different directories differ by up to
    0.25%, because the nimcache path reaches the generated C. Both arms must be
    built the same way; 93f7cc2 itself moves 0.23% between two builders.

+++ src/dingbat/gb/gb.nim
@@ -792,7 +792,14 @@ type
     # Both are the OBJ penalty algorithm's state; see tick_shifter's trigger.
     obj_penalty*:         int32
     obj_tile_fx*:         int32
-    m3_delay*:            int   # idle dots left at the head of mode 3
+    # Idle dots left at the head of mode 3 (the pipeline's lead over the CPU's
+    # register view; see M3_PIPE_DELAY in fifo_ppu). A byte, not an int, and
+    # for one reason: the mode 3 branch of fifo_tick_slow's dot loop asks
+    # "is the head spent?" once per M-cycle of mode 3 -- ~6,200 times a frame
+    # -- and a byte answers it in `ldrb`+`cbz` where a signed int needs
+    # `ldr`+`cmp`+`b.le`. The value is 0..12 by construction (M3_PIPE_MCYCLES
+    # * 4 + M3_PIPE_DELAY).
+    m3_delay*:            uint8
     # How far the pipeline lags the CPU's view of the PPU registers on THIS
     # line, in dots. Latched at the mode 2 -> 3 edge because the CPU M-cycle it
     # is derived from is 4 dots at normal speed and 2 in double speed. See
